+++ app/prefix_test/prefix_test.go
@@ -22,6 +22,7 @@ package prefix_test
 import (
 	"testing"
 
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -29,55 +30,131 @@ import (
 	"github.com/provenance-io/provenance/app"
 )
 
-func TestFullBIP44Path(t *testing.T) {
+func TestSetConfig(t *testing.T) {
 	cases := []struct {
-		name         string
-		expected     string
-		isTestnet    bool
-		seal         bool
-		panicMessage string
+		name        string
+		isTestnet   bool
+		seal        bool
+		expPath     string
+		expHRP      string
+		expCoinType int
+		expPanic    string
 	}{
 		{
-			// Note: The way our SetConfig function works, calling it with isTestnet = true makes some changes that aren't undone
-			// when later calling it with isTestnet = false. So to test the mainnet BIP44, we need to do it first.
-			name:         "has correct bip44th path for mainnet",
-			expected:     "m/44'/505'/0'/0/0",
-			isTestnet:    false,
-			seal:         false,
-			panicMessage: "",
+			name:        "mainnet",
+			isTestnet:   false,
+			seal:        false,
+			expPath:     "m/44'/505'/0'/0/0",
+			expHRP:      app.AccountAddressPrefixMainNet,
+			expCoinType: app.CoinTypeMainNet,
 		},
 		{
-			// Note: This is the 2nd to last test, so we're doing seal = true.
-			name:         "has correct bip44th path for testnet",
-			expected:     "m/44'/1'/0'/0/0",
-			isTestnet:    true,
-			seal:         true,
-			panicMessage: "",
+			name:        "testnet",
+			isTestnet:   true,
+			seal:        false,
+			expPath:     "m/44'/1'/0'/0/0",
+			expHRP:      app.AccountAddressPrefixTestNet,
+			expCoinType: app.CoinTypeTestNet,
 		},
 		{
-			// Note: The previous test should have had seal = true, making this an
+			name:        "back to mainnet",
+			isTestnet:   false,
+			seal:        false,
+			expPath:     "m/44'/505'/0'/0/0",
+			expHRP:      app.AccountAddressPrefixMainNet,
+			expCoinType: app.CoinTypeMainNet,
+		},
+		{
+			// This is the last valid test, so we're doing seal = true.
+			name:        "back to testnet with seal",
+			isTestnet:   true,
+			seal:        true,
+			expPath:     "m/44'/1'/0'/0/0",
+			expHRP:      app.AccountAddressPrefixTestNet,
+			expCoinType: app.CoinTypeTestNet,
+		},
+		{
+			// Note: A previous test should have had seal = true, making this an
 			// attempt to change the config after sealing it.
-			name:         "cannot double seal",
-			expected:     "",
-			isTestnet:    false,
-			seal:         true,
-			panicMessage: "Config is sealed",
+			name:      "already sealed: mainnet, no reseal",
+			isTestnet: false,
+			seal:      false,
+			expPanic:  "Config is sealed",
+		},
+		{
+			// Note: A previous test should have had seal = true, making this an
+			// attempt to change the config after sealing it.
+			name:      "already sealed: testnet, no reseal",
+			isTestnet: true,
+			seal:      false,
+			expPanic:  "Config is sealed",
+		},
+		{
+			// Note: A previous test should have had seal = true, making this an
+			// attempt to change the config after sealing it.
+			name:      "already sealed: mainnet, with reseal",
+			isTestnet: false,
+			seal:      true,
+			expPanic:  "Config is sealed",
+		},
+		{
+			// Note: A previous test should have had seal = true, making this an
+			// attempt to change the config after sealing it.
+			name:      "already sealed: testnet, with reseal",
+			isTestnet: true,
+			seal:      true,
+			expPanic:  "Config is sealed",
 		},
 	}
 
 	for _, tc := range cases {
 		t.Run(tc.name, func(t *testing.T) {
-			if len(tc.panicMessage) > 0 {
-				require.PanicsWithValue(t, tc.panicMessage, func() {
-					app.SetConfig(tc.isTestnet, tc.seal)
-				}, "SetConfig")
+			testFunc := func() {
+				app.SetConfig(tc.isTestnet, tc.seal)
+			}
+			if len(tc.expPanic) > 0 {
+				require.PanicsWithValue(t, tc.expPanic, testFunc, "SetConfig")
 			} else {
-				require.NotPanics(t, func() {
-					app.SetConfig(tc.isTestnet, tc.seal)
-				}, "SetConfig")
-				config := sdk.GetConfig()
-				fullBIP44Path := config.GetFullBIP44Path()
-				require.Equal(t, tc.expected, fullBIP44Path, "GetFullBIP44Path")
+				require.NotPanics(t, testFunc, "SetConfig")
+
+				expAddrPre := tc.expHRP
+				expValPre := tc.expHRP + "valoper"
+				expValPubPre := tc.expHRP + "valoperpub"
+				expAddrPrePub := tc.expHRP + "pub"
+				expConsPre := tc.expHRP + "valcons"
+				expConsPubPre := tc.expHRP + "valconspub"
+
+				sdkConfig := sdk.GetConfig()
+				fullBIP44Path := sdkConfig.GetFullBIP44Path()
+				addrPre := sdkConfig.GetBech32AccountAddrPrefix()
+				addrPubPre := sdkConfig.GetBech32AccountPubPrefix()
+				valPre := sdkConfig.GetBech32ValidatorAddrPrefix()
+				valPubPre := sdkConfig.GetBech32ValidatorPubPrefix()
+				consPre := sdkConfig.GetBech32ConsensusAddrPrefix()
+				consPubPre := sdkConfig.GetBech32ConsensusPubPrefix()
+				coinType := sdkConfig.GetCoinType()
+				purpose := sdkConfig.GetPurpose()
+
+				// Using require to check the main HRPs because if they're wrong, the rest will almost certainly also be wrong.
+				require.Equal(t, expAddrPre, app.AccountAddressPrefix, "AccountAddressPrefix")
+				require.Equal(t, expAddrPre, addrPre, "sdkConfig.GetBech32AccountAddrPrefix()")
+				// Asserts from here on out so we get a larger picture upon failure.
+				assert.Equal(t, expAddrPrePub, app.AccountPubKeyPrefix, "AccountPubKeyPrefix")
+				assert.Equal(t, expAddrPrePub, addrPubPre, "sdkConfig.GetBech32AccountPubPrefix()")
+				assert.Equal(t, expValPre, app.ValidatorAddressPrefix, "ValidatorAddressPrefix")
+				assert.Equal(t, expValPre, valPre, "sdkConfig.GetBech32ValidatorAddrPrefix()")
+				assert.Equal(t, expValPubPre, app.ValidatorPubKeyPrefix, "ValidatorPubKeyPrefix")
+				assert.Equal(t, expValPubPre, valPubPre, "sdkConfig.GetBech32ValidatorPubPrefix()")
+				assert.Equal(t, expConsPre, app.ConsNodeAddressPrefix, "ConsNodeAddressPrefix")
+				assert.Equal(t, expConsPre, consPre, "sdkConfig.GetBech32ConsensusAddrPrefix()")
+				assert.Equal(t, expConsPubPre, app.ConsNodePubKeyPrefix, "ConsNodePubKeyPrefix")
+				assert.Equal(t, expConsPubPre, consPubPre, "sdkConfig.GetBech32ConsensusPubPrefix()")
+
+				assert.Equal(t, tc.expPath, fullBIP44Path, "sdkConfig.GetFullBIP44Path()")
+				assert.Equal(t, tc.expCoinType, app.CoinType, "CoinType")
+				assert.Equal(t, tc.expCoinType, int(coinType), "sdkConfig.GetCoinType()")
+				assert.Equal(t, 44, app.Purpose, "Purpose")
+				assert.Equal(t, 44, int(purpose), "sdkConfig.GetPurpose()")
 			}
 		})
 	}