+++ app/upgrades_test.go
@@ -367,8 +367,8 @@ func (s *UpgradeTestSuite) TestUmberRC1() {
 	expInLog := []string{
 		"INF Pruning expired consensus states for IBC.",
 		"INF Done pruning expired consensus states for IBC.",
-		"INF Migrating legacy params.",
-		"INF Done migrating legacy params.",
+		"INF Migrating consensus params.",
+		"INF Done migrating consensus params.",
 		"INF Migrating bank params.",
 		"INF Done migrating bank params.",
 		"INF Migrating attribute params.",
@@ -400,8 +400,8 @@ func (s *UpgradeTestSuite) TestUmber() {
 	expInLog := []string{
 		"INF Pruning expired consensus states for IBC.",
 		"INF Done pruning expired consensus states for IBC.",
-		"INF Migrating legacy params.",
-		"INF Done migrating legacy params.",
+		"INF Migrating consensus params.",
+		"INF Done migrating consensus params.",
 		"INF Migrating bank params.",
 		"INF Done migrating bank params.",
 		"INF Migrating attribute params.",